requirements.txt
requests
python-pptx
customtkinter
